[10-refactoring] LotationsRes -> PBRes.kt 로 이름 변경
