+++ app/src/main/java/com/example/pozi_android/data/remote/network/PBRes.kt
@@ -3,7 +3,7 @@ package com.example.pozi_android.data.remote.network
 import com.example.pozi_android.data.remote.spec.PBResponce
 import com.google.gson.annotations.SerializedName
 
-class LocationRes {
+class PBRes {
     @SerializedName("items")
     val locations: List<PBResponce> = listOf()
 }
