+++ app/src/main/java/com/example/pozi_android/data/repository/api/ServiceRepositoryImpl.kt
@@ -1,10 +1,9 @@
-package com.example.pozi_android.domain.repository
+package com.example.pozi_android.data.repository.api
 
-import android.util.Log
 import com.example.pozi_android.data.remote.network.DataResult
-import com.example.pozi_android.data.remote.network.LocationRes
+import com.example.pozi_android.data.remote.network.PBRes
 import com.example.pozi_android.data.remote.network.RetrofitInterface
-import com.example.pozi_android.data.repository.api.ServiceRepository
+import com.example.pozi_android.domain.repository.ServiceRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 
@@ -13,7 +12,7 @@ class ServiceRepositoryImpl(
     private val ioDispatcher: CoroutineDispatcher
 ) : ServiceRepository {
 
-    override suspend fun getPhotoBoothList(): DataResult<LocationRes> = withContext(ioDispatcher) {
+    override suspend fun getPBList(): DataResult<PBRes> = withContext(ioDispatcher) {
         try {
             val response = api.getPhotoBoothList()
             when {
