+++ app/src/main/java/com/example/pozi_android/di/RepositoryModule.kt
@@ -1,8 +1,8 @@
 package com.example.pozi_android.di
 
 import com.example.pozi_android.data.remote.network.RetrofitInterface
-import com.example.pozi_android.data.repository.api.ServiceRepository
-import com.example.pozi_android.domain.repository.ServiceRepositoryImpl
+import com.example.pozi_android.domain.repository.ServiceRepository
+import com.example.pozi_android.data.repository.api.ServiceRepositoryImpl
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
