+++ app/src/main/java/com/example/pozi_android/domain/entity/PB.kt
@@ -1,0 +1,9 @@
+package com.example.pozi_android.domain.entity
+
+data class PB(
+    val id: Int,
+    val address: String,
+    val lat: Double,
+    val lng: Double,
+    val name: String
+)
