--- app/src/main/java/com/example/pozi_android/domain/mapper/PBinfoMapper.kt
@@ -1,7 +1,0 @@
-package com.example.pozi_android.domain.mapper
-
-import com.example.pozi_android.data.remote.spec.PBResponce
-
-//object PBinfoMapper {
-//    fun mapperToPB(PBResponce : PBResponce) :
-//}
