--- app/src/main/java/com/example/pozi_android/domain/mapper/empty.kt
@@ -1,4 +1,0 @@
-package com.example.pozi_android.domain.mapper
-
-object empty {
-}
