+++ app/src/main/java/com/example/pozi_android/domain/repository/ServiceRepository.kt
@@ -1,0 +1,8 @@
+package com.example.pozi_android.domain.repository
+
+import com.example.pozi_android.data.remote.network.DataResult
+import com.example.pozi_android.data.remote.network.PBRes
+
+interface ServiceRepository {
+    suspend fun getPBList(): DataResult<PBRes>
+}
