+++ app/src/main/java/com/example/pozi_android/ui/main/MainViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.example.pozi_android.data.remote.network.DataResult
-import com.example.pozi_android.data.remote.network.LocationRes
-import com.example.pozi_android.data.repository.api.ServiceRepository
+import com.example.pozi_android.data.remote.network.PBRes
+import com.example.pozi_android.domain.repository.ServiceRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
@@ -16,13 +16,13 @@ import javax.inject.Inject
 class MainViewModel @Inject constructor(private val serviceRepository: ServiceRepository) :
     ViewModel() {
 
-    private val _photoBoothList: MutableLiveData<DataResult<LocationRes>> =
+    private val _photoBoothList: MutableLiveData<DataResult<PBRes>> =
         MutableLiveData()
-    val photoBoothList: LiveData<DataResult<LocationRes>> get() = _photoBoothList
+    val photoBoothList: LiveData<DataResult<PBRes>> get() = _photoBoothList
 
     fun getCenterList() {
         CoroutineScope(Dispatchers.IO).launch {
-            _photoBoothList.postValue(serviceRepository.getPhotoBoothList())
+            _photoBoothList.postValue(serviceRepository.getPBList())
         }
     }
